+++ Changelog.md
@@ -6,6 +6,7 @@ Released on ?.
 
 ✨ New Features / Improvements
 
+- Added importing .xctsk files as layers
 - Added the "Terrain slope direction" map overlay and showing a color legend
 - Display message band when terrain is not initialized yet
 - Using CesiumJS version 1.138.0